requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
channels==3.0.5
channels-redis==4.1.0
charset-normalizer==3.3.1
constantly==23.10.4
cryptography==41.0.5
daphne==3.0.2
defusedxml==0.8.0rc2
Django==4.2.6
django-allauth==0.58.0
django-appconf==1.0.5
django-extensions==3.2.3
django-ipware==5.0.2
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-redis==5.4.0
django-simple-captcha==0.5.20
djangorestframework==3.14.0
djangorestframework-recaptcha==0.2.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
google-auth==2.23.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Markdown==3.5.1
mongoengine==0.27.0
msgpack==1.0.7
oauthlib==3.2.2
path==16.7.1
Pillow==10.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymongo==4.5.0
pyOpenSSL==23.3.0
pyotp==2.9.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
service-identity==23.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.4.2
social-auth-storage-mongoengine==1.0.1
sqlparse==0.4.4
tokens==0.0.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.0.7
users==1.0.dev0
zope.interface==6.1
